+++ CHANGELOG.md
@@ -9,6 +9,11 @@ This Changelog tracks changes to this project. The notes below include a summary
 - `fixed` for any bug fixes.
 - `security` in case of vulnerabilities.
 
+## Version `0.26.3` - 20 Oct 2022
+
+- `added` current ctDNA and microbiome analysis files into the assay schemas
+- `removed` ctDNA and microbiome analysis schemas
+
 ## Version `0.26.2` - 10 Oct 2022
 
 - `added` MIBI assay support